Refactor auth excpetions to separate package
