+++ device-telemetry/app/com/microsoft/azure/iotsolutions/devicetelemetry/webservice/auth/exceptions/NotAuthorizedException.java
@@ -1,6 +1,6 @@
 // Copyright (c) Microsoft. All rights reserved.
 
-package com.microsoft.azure.iotsolutions.devicetelemetry.webservice.auth;
+package com.microsoft.azure.iotsolutions.devicetelemetry.webservice.auth.exceptions;
 
 /**
  * This exception is thrown when the user is not authorized to perform the action.
